Throw an error in hookgen if sdk refs aren't found.

+++ setup/Setup/HookGenTask.cs
@@ -4,16 +4,16 @@ using MonoMod.RuntimeDetour.HookGen;
 using System;
 using System.IO;
 using System.Windows;
-using static Terraria.ModLoader.Setup.Program;
 
 namespace Terraria.ModLoader.Setup
 {
 	internal class HookGenTask : SetupOperation
 	{
+		private const string dotnetSdkVersion = "6.0.0";
 		private const string libsPath = "src/tModLoader/Terraria/Libraries";
 		private const string binLibsPath = "src/tModLoader/Terraria/bin/Release/net6.0/Libraries";
 		private const string tmlAssemblyPath = @"src/tModLoader/Terraria/bin/Release/net6.0/tModLoader.dll";
-		private const string installedNetRefs = @"\dotnet\packs\Microsoft.NETCore.App.Ref\6.0.0\ref\net6.0";
+		private const string installedNetRefs = $@"\dotnet\packs\Microsoft.NETCore.App.Ref\{dotnetSdkVersion}\ref\net6.0";
 
 		public HookGenTask(ITaskInterface taskInterface) : base(taskInterface)
 		{
@@ -34,23 +34,38 @@ namespace Terraria.ModLoader.Setup
 
 			taskInterface.SetStatus($"Hooking: tModLoader.dll -> TerrariaHooks.dll");
 
-			HookGen(tmlAssemblyPath, outputPath);
+			if (!HookGen(tmlAssemblyPath, outputPath)) {
+				taskInterface.SetStatus("Cancelled");
+				return;
+			}
 
 			File.Delete(Path.ChangeExtension(outputPath, "pdb"));
 
 			MessageBox.Show("Success. Make sure you diff tModLoader after this");
 		}
 
-		public static void HookGen(string inputPath, string outputPath)
+		public static bool HookGen(string inputPath, string outputPath)
 		{
+			string dotnetReferencesDirectory = Environment.GetFolderPath(Environment.SpecialFolder.ProgramFiles) + installedNetRefs;
+
+			// Ensure that refs are present, for gods sake!
+			if (!Directory.Exists(dotnetReferencesDirectory) || Directory.GetFiles(dotnetReferencesDirectory, "*.dll").Length == 0) {
+				// Replace with exceptions if this is ever called in CLI.
+				MessageBox.Show(
+					$@"Unable to find reference libraries for .NET SDK '{dotnetSdkVersion}' - ""{dotnetReferencesDirectory}"" does not exist.",
+					$".NET SDK {dotnetSdkVersion} not found",
+					MessageBoxButton.OK
+				);
+
+				return false;
+			}
+
 			using var mm = new MonoModder {
 				InputPath = inputPath,
 				OutputPath = outputPath,
 				ReadingMode = ReadingMode.Deferred,
 
-				DependencyDirs = {
-					Environment.GetFolderPath(Environment.SpecialFolder.ProgramFiles) + installedNetRefs
-				},
+				DependencyDirs = { dotnetReferencesDirectory },
 				MissingDependencyThrow = false,
 			};
 
@@ -62,9 +77,14 @@ namespace Terraria.ModLoader.Setup
 			var gen = new HookGenerator(mm, "TerrariaHooks") {
 				HookPrivate = true,
 			};
+
 			gen.Generate();
+
 			RemoveModLoaderTypes(gen.OutputModule);
+
 			gen.OutputModule.Write(outputPath);
+
+			return true;
 		}
 
 		private static void RemoveModLoaderTypes(ModuleDefinition module)
